requirements.txt
fastapi~=0.111.1
uvicorn
pydantic~=2.8.2
neo4j~=5.22.0

langchain~=0.2.9
app~=0.0.1
starlette~=0.37.2
python-dotenv~=1.0.1
streamlit~=1.36.0
retry~=0.9.2
torch~=2.3.1
transformers~=4.42.4